Update Anna's bio + add support for odd numbers of contributors (#49)
* Add Anna to core team, update styling to allow for odd numbers of contributors
* Fix up some styling on window resize

I went to go make a small change to update my pic on the website and move to the core contributors section, but it turns out that the website doesn't really support odd #s of contributors yet, so made some CSS changes to move the borders / background colors to live on the individual sections.

Side effect: I un-rounded the corners on the contributors/funders sections because I wasn't sure how to best do rounding with an row with only one element.

Testing: Tested collapsing page down to min size, expanding to full screen. Visual check for regressions.